Refine an abstracted division by every quotient fact that fired
The seven facts about a quotient that came with the abstraction refinement
were the highest-firing seven of eighteen. The other eleven were left because
those seven turned out to be a wash on the family they were measured for, and
extending a set that does not pay wants a reason to expect the next one to.
The reason is narrower than "more facts are better": these eleven are what
the measurement named, and where one of them is the crux of a query it is
worth an exact 128-bit divider. Below, five queries that took between three
seconds and forever now take under a second and a half.

By firing count over the 73 Certora files STP cannot decide, the eleven are:

  x >=u ((t << 1) >> (t << s))    125    t >=u ((x >> s) << 1)          14
  t != -(s & ~x)                   62    x >=u ((x | t) & (s << 1))     10
  x >=u ((s >> (s << t)) << 1)     54    (x & -t) >=u (s & t)            9
  x != t + t + (x | s)             26    x >=u (t ^ (t >> (s >> 1)))     3
  (x >> t) != (s | t)               3    x >=u ((x | s) & (t << 1))      2
  x >=u (s ^ (s >> (t >> 1)))       2

Fifteen more exist upstream and are not here, having fired nothing. That is
real evidence rather than an artefact of the counter: every fact the source
installs is permanent and it stops at the first one a candidate breaks, so
"fired nothing" says no candidate on that corpus broke the fact without also
breaking something earlier. It is not proof of uselessness -- a stronger fact
installed sooner would exclude more of what comes after it -- and one of the
fifteen is worth a second look on exactly that ground. UDIV33,
`(s ^ (x | t)) >=u (t ^ 1)`, rules out 49% of the (x, s, t) cube at six bits,
which is more than every fact here except UDIV7 at 58%, and it needs no
variable shift. It sits at position 27 of 30 in the source's order, which is
the whole of why it never fired. That is a separate experiment.

Two of the eleven shift left by a variable amount, which STP could not do --
BBShiftRightByVariable is right-shift only. BBShiftLeftByVariable is its
mirror: the same multiplexer per stage with the constant shift going the
other way, and the same saturating tail, since SMT-LIB's bvshl and bvlshr
agree that an amount at or past the width clears the vector rather than
leaving the value alone.

One fact is not true at every width. `x != t + t + (x | s)` is false at one
bit, where x = 1 over s = 0 gives the totalised quotient 1 and the right-hand
side is 1 as well. The solver these come from sidesteps every such
restriction by refusing to abstract anything narrower than three bits; STP's
threshold is a flag whose default is 64 but whose floor is one, so the
minimum is carried per fact instead and the chooser skips a fact the
abstraction in front of it is too narrow for. The minimum is exact rather
than cautious, and a test says so by checking the fact really does fail one
bit below it.

What this buys, on the queries where one fact is the crux -- five are pinned,
each unsatisfiable for exactly the reason one fact supplies, median of three
runs at a 60s cap:

  x >=u ((t << 1) >> (t << s))    160 bits   8.85s -> 1.10s
  t >=u ((x >> s) << 1)           128 bits   7.00s -> 0.55s
  x >=u ((x | s) & (t << 1))      128 bits   did not finish -> 0.29s
  (x & -t) >=u (s & t)            128 bits   3.32s -> 0.34s
  x >=u (t ^ (t >> (s >> 1)))     128 bits   3.85s -> 0.46s

In aggregate, on the family these were measured for, it is a wash -- which is
what the first seven facts were, and it is reported the same way. A paired run
over all 1029 `20241113-Certora` files, both binaries on the same file back to
back at 30s each so the two see the same machine: no disagreement anywhere
both answered, 639 solved before and 641 after, 11 gained against 9 lost, PAR2
30182 against 30080. This family's run-to-run spread is 13 files, so nothing
there could be anything but a wash.

The counters say where it went. Over the 630 files both binaries solved, the
two take byte-identical paths on 602 of them -- same SAT calls, same rounds,
same lemmas -- so on those no new fact was ever the first thing a candidate
broke. It is not that the abstraction never sees a division: a 60-file census
of the solved ones finds 23 that abstract at least one. It is that the
schemas and bounds already there catch the candidate first.

On the 28 files where the new facts do fire, 80 of them are installed in
place of 23 blocking lemmas, and individual files move a long way in both
directions:

  0672.smt2   41 SAT calls -> 35,  96 blocking -> 61,  29.63s -> 2.83s
  1671.smt2   63 SAT calls -> 20, 100 blocking -> 42,  20.38s -> 11.47s
  2029.smt2   38 SAT calls -> 68,   5 blocking -> 43,   8.08s -> 19.86s
  1898.smt2   16 SAT calls -> 52,   1 blocking -> 38,   9.16s -> 16.16s

The two that got worse are the same mechanism as the two that got better. A
fact installed early is a clause the search did not have before, and where
that sends the search afterwards is not something the fact can be credited
with or blamed for. Summed over the 630: SAT calls +0.3%, refinement rounds
+0.4%, blocking lemmas -1.3%, wall time -0.3%.

Nothing here trusts the transcription, and the checks are the ones the first
seven brought, widened. Each fact is evaluated at the true quotient over
every pair of operands at every width from its declared minimum to five, so
one that is false of division dies whatever its circuit does and a minimum
that is wrong in either direction is caught. Each must refute some triple.
And the circuit that reaches the solver is asked, over all 4096 triples at
four bits and all 512 at three, whether it permits that triple, which must
match its predicate exactly -- three bits as well as four because a width
that is not a power of two is where a barrel shifter's stages stop dividing
the width and the saturating tail has to make up the difference.

That is 82944 solves against the seven facts' 28672, and it runs in 0.3s
rather than 28s: the circuit is now built once per fact and the triple asked
for by assumption, where before each triple paid for its own blast, AIG
rewrite and CNF derivation. The facts come from the refiner's own table
rather than a list beside it, so a fact added to one and not the other is not
a fact that goes untested.

+++ include/stp/ToSat/BVAbstractionRefiner.h
@@ -151,7 +151,7 @@ enum
 DLL_PUBLIC MulSchemaChoice chooseMulSchema(const std::vector<bool>& aBits,
                                            const std::vector<bool>& bBits,
                                            const std::vector<bool>& tBits,
-                                           unsigned installedSchemas);
+                                           uint64_t installedSchemas);
 
 // The algebraic facts an abstracted BVDIV or BVMOD is refined with.
 //
@@ -213,15 +213,19 @@ enum
   DIV_SCHEMA_INSTALLED_REMAINDER_AT_MOST_DIVIDEND = 8u,
   DIV_SCHEMA_INSTALLED_REMAINDER_BELOW_DIVISOR = 16u,
   DIV_SCHEMA_INSTALLED_QUOTIENT_AT_MOST_DIVIDEND = 32u,
-  // ... and one apiece for the seven DivLemma facts, which are
-  // unconditional for the same reason and tracked the same way. The first
-  // of them is 64; `DIV_LEMMA_INSTALLED(i)` is the bit for the i'th.
+  // ... and one apiece for the DivLemma facts, which are unconditional for
+  // the same reason and tracked the same way. The first of them is 64;
+  // `divLemmaInstalledBit(i)` is the bit for the i'th.
   DIV_LEMMA_INSTALLED_FIRST = 64u
 };
 
-inline unsigned divLemmaInstalledBit(unsigned index)
+// Sixty-four bits and not thirty-two. The field is nowhere near full, but
+// the shift below is what would overflow -- silently, and only once the
+// table grew past twenty-six entries, which is a thing a later commit does
+// without ever looking here.
+inline uint64_t divLemmaInstalledBit(unsigned index)
 {
-  return DIV_LEMMA_INSTALLED_FIRST << index;
+  return (uint64_t)DIV_LEMMA_INSTALLED_FIRST << index;
 }
 
 // The DivLemma facts the chooser offers, in the order it offers them, and
@@ -265,7 +269,7 @@ DLL_PUBLIC DivSchemaChoice chooseDivSchema(Kind opKind,
                                            const std::vector<bool>& aBits,
                                            const std::vector<bool>& bBits,
                                            const std::vector<bool>& tBits,
-                                           unsigned installedSchemas);
+                                           uint64_t installedSchemas);
 
 // A variable that holds exactly when `lv <= rv`. Shared by the comparison
 // refinement, which is where it comes from, and by the division bounds,
@@ -335,7 +339,7 @@ struct BVTermAbstraction
   // each one.
   unsigned schemaRounds = 0;
   // Which of the unconditional schemas are already in the solver.
-  unsigned installedSchemas = 0;
+  uint64_t installedSchemas = 0;
   // How far up the exact encoding has been pushed, for an escalation that
   // goes a piece at a time; see bv_term_abstraction_inc_bitblast. Zero
   // until the first piece, and equal to the width once `defined` is set.
